Fixes compiler warnigns

+++ cocos2d/ccTypes.h
@@ -141,9 +141,8 @@ static inline BOOL ccc4FEqual(ccColor4F a, ccColor4F b)
 	
 static inline ccColor4B ccc4BFromccc4F(ccColor4F c)
 {
-	return (ccColor4B){c.r*255.f, c.g*255.f, c.b*255.f, c.a*255.f};
+	return (ccColor4B){(GLubyte)(c.r*255), (GLubyte)(c.g*255), (GLubyte)(c.b*255), (GLubyte)(c.a*255)};
 }
-	
 
 
 /** A vertex composed of 2 GLfloats: x, y
